fix(storage): reject non-string decision keys; split the availability tests
`AvailabilityLedger.from_dict` coerced mapping keys with `str()`, so
`{1: {"status": "failed"}, "1": {"status": "present"}}` — which a YAML loader
or a Python adapter can produce — collapsed to one entry, with iteration order
deciding which record survived. Reversing it flipped `for_family("1")` from
non-comparable to comparable, so a discarded `FAILED` record could license a
conclusion. Family names and both override key halves are now rejected rather
than coerced.

This is the same rule `canonical_form` already applies to mapping keys, in the
one place that had not adopted it. The two modules are leaves and share
nothing by design, so the rule is restated rather than imported — which is the
drift this branch keeps finding, so both sites now carry a test instead of a
comment promising they agree. `versioning`'s `section_schema_versions` keeps
its `str()` deliberately: it is one of the five informational axes, which
parse defensively because no decision reads them. That distinction is now
pinned by a test rather than left implicit.

The new tests pushed `test_availability.py` past this repo's 1200-line test
cap, so it is split by subject: `test_availability_documents.py` takes the
`from_dict`/`to_dict` boundary contracts — which malformed inputs are refused,
and whether a written document describes the object it came from — and the
original keeps the in-memory contracts. Several of the document tests exist
because a serializer disagreed with its own reader, so grouping them makes
that comparison easy to see.

Regression tests verified failing against the pre-fix parse (12 of 14).
